+++ src/__init__.py
@@ -1,8 +1,8 @@
 """
 Emby Watch Party - Source Modules
-Version: 1.6.4
+Version: 1.6.5
 Codename: Synced Popcorn
 """
 
-__version__ = "1.6.4"
+__version__ = "1.6.5"
 __codename__ = "Synced Popcorn"